+++ src/lib/lib.rs
@@ -101,6 +101,7 @@ extern crate mustache;
 extern crate chrono;
 extern crate uuid;
 extern crate yaml_rust;
+extern crate mime_guess;
 
 pub use parser::Parser;
 pub use book::Book;